pyproject.toml
[project]
name = "delphi-press"
dynamic = ["version"]
description = "Media headline forecasting with multi-agent Delphi pipeline"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.0",
    "openai>=1.0",
    "jinja2>=3.1",
    "httpx>=0.27",
    "feedparser>=6.0",
    "fastfeedparser>=0.5.7",
    "trafilatura>=1.12",
    "aiosqlite>=0.20",
    "rapidfuzz>=3.0",
    "numpy>=1.26",
    "scikit-learn>=1.4",
    "scipy>=1.13",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "arq>=0.26",
    "sse-starlette>=2.0",
    "redis[hiredis]>=5.0",
    "cryptography>=43.0",
    "PyJWT>=2.9",
    "bcrypt>=4.2",
    "email-validator>=2.0",
    "python-multipart>=0.0.22",
    "pyarrow>=15.0",
    "claude-agent-sdk>=0.1.58",
]

[project.optional-dependencies]
ml = [
    "hdbscan>=0.8.38",
]
eval = [
    "statsmodels>=0.14",
    "duckdb>=1.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.per-file-ignores]
"src/llm/prompts/**/*.py" = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "mkdocs-copy-to-llm>=0.2.10",
    "mkdocs-include-markdown-plugin>=7.2.2",
    "mkdocs-llmstxt>=0.5.0",
    "mkdocs-material",
    "pymdown-extensions",
]
